requirements.txt
fastapi
uvicorn
pypdf
chromadb
langchain
langchain-community
langchain-groq
python-dotenv
sentence-transformers
langchain-huggingface
